fix(crc32): properly check for crc32 by looking for __CRC32__ instead of __SSE4_2__
sse4 is supposed to imply crc32, so the current check should be fine. but checking for crc32 is better because it's what's actually needed

also, zig has a bug where it does not properly detect crc32 support on some platforms supporting sse4 (e.g. tiger lake and arrow lake). clang avoids this by explicitly adding crc32 in the code if sse4 is supported. zig does not do this, so we get compilation errors here when building with zig. this fix will at least prevent the build errors